refactor(rabbitmq): split catalog and client boundaries
